### 2020-02-02 2b2e055
Including shell of spcaWrapper function
