+++ inst/REFERENCES.bib
@@ -32,6 +32,14 @@
   publisher={BioMed Central}
 }
 
+@article{erichson2018sparse,
+  title={Sparse Principal Component Analysis via Variable Projection},
+  author={Erichson, N. Benjamin and Zeng, Peng and Manohar, Krithika and Brunton, Steven L. and Kutz, J. Nathan and Aravkin, Aleksandr Y.},
+  journal={ArXiv},
+  year={2018},
+  volume={abs/1804.00341}
+}
+
 @article {boileau2020,
 	author = {Boileau, Philippe and Hejazi, Nima S. and Dudoit, Sandrine},
 	title = {Exploring High-Dimensional Biological Data with Sparse Contrastive Principal Component Analysis},
