### 2020-02-02 886399f
completed draft of spcaWrapper function
