+++ NAMESPACE
@@ -11,6 +11,7 @@ importFrom(cluster,pam)
 importFrom(cluster,silhouette)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,full_join)
+importFrom(elasticnet,rootmatrix)
 importFrom(elasticnet,spca)
 importFrom(kernlab,as.kernelMatrix)
 importFrom(kernlab,specc)
@@ -21,6 +22,8 @@ importFrom(origami,make_folds)
 importFrom(origami,training)
 importFrom(origami,validation)
 importFrom(purrr,reduce)
+importFrom(sparsepca,rspca)
+importFrom(sparsepca,spca)
 importFrom(stats,cov)
 importFrom(stats,cutree)
 importFrom(stats,dist)
