### 2020-02-02 0061667
Updated spcaWrapper function to handle zero cov matrices
