### 2020-02-02 73a2c7b
updating documentation for fitGrid and bpFitGrid
