+++ R/fitGrid.R
@@ -17,6 +17,12 @@
 #' @param c_contrasts A \code{list} of contrastive covariances.
 #' @param contrasts A \code{numeric} vector of the contrastive parameters used
 #'  to compute the contrastive covariances.
+#' @param alg A \code{character} indicating the SPCA algorithm used to sparsify
+#'   the contrastive loadings. Currently supports \code{iterative} for the
+#'   \insertRef{zou2006sparse}{scPCA} implemententation, \code{var_proj} for the
+#'   non-randomized \insertRef{erichson2018sparse}{scPCA} solution, and
+#'   \code{rand_var_proj} fir the randomized
+#'   \insertRef{erichson2018sparse}{scPCA} result.
 #' @param penalties A \code{numeric} vector of the penalty terms.
 #' @param n_eigen A \code{numeric} indicating the number of eigenvectors to be
 #'  computed.
@@ -43,7 +49,8 @@
 #'     \item penalty - the optimal L1 penalty term
 #'   }
 #'
-#' @importFrom elasticnet spca
+#' @importFrom sparsepca spca
+#' @importFrom elasticnet rootmatrix
 #' @importFrom stats kmeans dist hclust cutree
 #' @importFrom cluster pam silhouette
 #'
@@ -72,12 +79,12 @@ fitGrid <- function(target, target_valid = NULL, center, scale,
               symmetric = TRUE
             )$vectors[, seq_len(n_eigen)]
           } else {
-            res <- elasticnet::spca(c_contrasts[[x]],
-              K = n_eigen,
-              para = rep(y, n_eigen),
-              type = "Gram",
-              sparse = "penalty"
-            )$loadings
+            res <- spcaWrapper(
+              alg = alg,
+              contrast_cov = c_contrasts[[x]],
+              k = n_eigen,
+              penalty = y
+            )
           }
           colnames(res) <- paste0("V", as.character(seq_len(n_eigen)))
           res
@@ -212,6 +219,12 @@ fitGrid <- function(target, target_valid = NULL, center, scale,
 #' @param c_contrasts A \code{list} of contrastive covariances.
 #' @param contrasts A \code{numeric} vector of the contrastive parameters used
 #'  to compute the contrastive covariances.
+#' @param alg A \code{character} indicating the SPCA algorithm used to sparsify
+#'  the contrastive loadings. Currently supports \code{iterative} for the
+#'  \insertRef{zou2006sparse}{scPCA} implemententation, \code{var_proj} for the
+#'  non-randomized \insertRef{erichson2018sparse}{scPCA} solution, and
+#'  \code{rand_var_proj} fir the randomized
+#'  \insertRef{erichson2018sparse}{scPCA} result.
 #' @param penalties A \code{numeric} vector of the penalty terms.
 #' @param n_eigen A \code{numeric} indicating the number of eigenvectors to be
 #'  computed.
@@ -238,7 +251,8 @@ fitGrid <- function(target, target_valid = NULL, center, scale,
 #'     \item penalty - the optimal L1 penalty term
 #'   }
 #'
-#' @importFrom elasticnet spca
+#' @importFrom elasticnet rootmatrix
+#' @importFrom sparsepca spca
 #' @importFrom stats kmeans dist hclust cutree
 #' @importFrom cluster pam silhouette
 #' @importFrom BiocParallel bplapply
@@ -268,12 +282,12 @@ bpFitGrid <- function(target, target_valid = NULL, center, scale,
               symmetric = TRUE
             )$vectors[, seq_len(n_eigen)]
           } else {
-            res <- elasticnet::spca(c_contrasts[[x]],
-              K = n_eigen,
-              para = rep(y, n_eigen),
-              type = "Gram",
-              sparse = "penalty"
[…]
