### 2020-02-02 17c2a1c
passes all old tests, updated docs
