+++ tests/testthat/test-fitGrid.R
@@ -16,7 +16,7 @@ fit <- fitGrid(
   target = target, center = center, scale = scale,
   c_contrasts = c_contrasts, contrasts = contrasts,
   penalties = penalties, n_eigen = n_eigen, n_centers = n_centers,
-  clust_method = "kmeans"
+  clust_method = "kmeans", alg = "iterative"
 )
 
 test_that("The output is a list of length 4", {
@@ -47,62 +47,64 @@ test_that("All cluster methods (kmeans, pam, hclust) run without errors", {
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
-    n_centers = n_centers, clust_method = "kmeans"
+    n_centers = n_centers, clust_method = "kmeans",
+    alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
-    n_centers = n_centers, clust_method = "pam"
-  ))
-  expect_silent(fitGrid(
-    target = target, center = center, scale = scale,
-    c_contrasts = c_contrasts, contrasts = contrasts,
-    penalties = penalties, n_eigen = n_eigen,
-    n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "complete"
-  ))
-  expect_silent(fitGrid(
-    target = target, center = center, scale = scale,
-    c_contrasts = c_contrasts, contrasts = contrasts,
-    penalties = penalties, n_eigen = n_eigen,
-    n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "average"
+    n_centers = n_centers, clust_method = "pam",
+    alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
     n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "single"
+    linkage_method = "complete", alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
     n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "ward.D"
+    linkage_method = "average", alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
     n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "mcquitty"
+    linkage_method = "single", alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
     n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "median"
+    linkage_method = "ward.D", alg = "iterative"
   ))
   expect_silent(fitGrid(
     target = target, center = center, scale = scale,
     c_contrasts = c_contrasts, contrasts = contrasts,
     penalties = penalties, n_eigen = n_eigen,
     n_centers = n_centers, clust_method = "hclust",
-    linkage_method = "centroid"
+    linkage_method = "mcquitty", alg = "iterative"
+  ))
+  expect_silent(fitGrid(
[…]
