+++ R/fitGrid.R
@@ -21,7 +21,7 @@
 #'   the contrastive loadings. Currently supports \code{iterative} for the
 #'   \insertRef{zou2006sparse}{scPCA} implemententation, \code{var_proj} for the
 #'   non-randomized \insertRef{erichson2018sparse}{scPCA} solution, and
-#'   \code{rand_var_proj} fir the randomized
+#'   \code{rand_var_proj} for the randomized
 #'   \insertRef{erichson2018sparse}{scPCA} result.
 #' @param penalties A \code{numeric} vector of the penalty terms.
 #' @param n_eigen A \code{numeric} indicating the number of eigenvectors to be
@@ -49,14 +49,12 @@
 #'     \item penalty - the optimal L1 penalty term
 #'   }
 #'
-#' @importFrom sparsepca spca
-#' @importFrom elasticnet rootmatrix
 #' @importFrom stats kmeans dist hclust cutree
 #' @importFrom cluster pam silhouette
 #'
 #' @keywords internal
 fitGrid <- function(target, target_valid = NULL, center, scale,
-                    c_contrasts, contrasts, penalties, n_eigen,
+                    c_contrasts, contrasts, alg, penalties, n_eigen,
                     clust_method = c("kmeans", "pam", "hclust"),
                     n_centers, max_iter = 10,
                     linkage_method = "complete") {
@@ -251,8 +249,6 @@ fitGrid <- function(target, target_valid = NULL, center, scale,
 #'     \item penalty - the optimal L1 penalty term
 #'   }
 #'
-#' @importFrom elasticnet rootmatrix
-#' @importFrom sparsepca spca
 #' @importFrom stats kmeans dist hclust cutree
 #' @importFrom cluster pam silhouette
 #' @importFrom BiocParallel bplapply
@@ -260,7 +256,7 @@ fitGrid <- function(target, target_valid = NULL, center, scale,
 #' @keywords internal
 bpFitGrid <- function(target, target_valid = NULL, center, scale,
                       c_contrasts, contrasts, penalties, n_eigen,
-                      clust_method = c("kmeans", "pam", "hclust"),
+                      alg, clust_method = c("kmeans", "pam", "hclust"),
                       n_centers, max_iter = 10,
                       linkage_method = "complete") {
   # preliminaries

[…]
