### 2020-02-02 d0a0eb3
Adding tests for new spca optimizers
