+++ tests/testthat/test-fitGrid.R
@@ -12,34 +12,63 @@ c_contrasts <- contrastiveCov(target, background, contrasts, TRUE, TRUE)
 penalties <- c(0, 0.1, 1)
 n_eigen <- 2
 n_centers <- 4
-fit <- fitGrid(
+fit_iterative <- fitGrid(
   target = target, center = center, scale = scale,
   c_contrasts = c_contrasts, contrasts = contrasts,
   penalties = penalties, n_eigen = n_eigen, n_centers = n_centers,
   clust_method = "kmeans", alg = "iterative"
 )
+fit_var_proj <- fitGrid(
+  target = target, center = center, scale = scale,
+  c_contrasts = c_contrasts, contrasts = contrasts,
+  penalties = penalties, n_eigen = n_eigen, n_centers = n_centers,
+  clust_method = "kmeans", alg = "var_proj"
+)
+set.seed(163)
+fit_rand_var_proj <- fitGrid(
+  target = target, center = center, scale = scale,
+  c_contrasts = c_contrasts, contrasts = contrasts,
+  penalties = penalties, n_eigen = n_eigen, n_centers = n_centers,
+  clust_method = "kmeans", alg = "rand_var_proj"
+)
 
 test_that("The output is a list of length 4", {
-  expect_equal(length(fit), 5)
+  expect_equal(length(fit_iterative), 5)
+  expect_equal(length(fit_var_proj), 5)
+  expect_equal(length(fit_rand_var_proj), 5)
 })
 
 test_that("The rotation matrix is ncol(target) by n_eigen", {
-  expect_equal(nrow(fit$rotation[[1]]), ncol(target))
-  expect_equal(ncol(fit$rotation[[1]]), n_eigen)
+  expect_equal(nrow(fit_iterative$rotation[[1]]), ncol(target))
+  expect_equal(ncol(fit_iterative$rotation[[1]]), n_eigen)
+  expect_equal(nrow(fit_var_proj$rotation[[1]]), ncol(target))
+  expect_equal(ncol(fit_var_proj$rotation[[1]]), n_eigen)
+  expect_equal(nrow(fit_rand_var_proj$rotation[[1]]), ncol(target))
+  expect_equal(ncol(fit_rand_var_proj$rotation[[1]]), n_eigen)
 })
 
 test_that("The projection matrix is nrow(target) by n_eigen", {
-  expect_equal(nrow(fit$x[[2]]), nrow(target))
-  expect_equal(ncol(fit$x[[2]]), n_eigen)
+  expect_equal(nrow(fit_iterative$x[[2]]), nrow(target))
+  expect_equal(ncol(fit_iterative$x[[2]]), n_eigen)
+  expect_equal(nrow(fit_var_proj$x[[2]]), nrow(target))
+  expect_equal(ncol(fit_var_proj$x[[2]]), n_eigen)
+  expect_equal(nrow(fit_rand_var_proj$x[[2]]), nrow(target))
+  expect_equal(ncol(fit_rand_var_proj$x[[2]]), n_eigen)
 })
 
 test_that(paste(
   "The optimal contrast and penalty terms are real-valued",
   "and non-negative"
 ), {
-  idx <- which.max(fit$ave_sil_widths)
-  expect_gte(fit$contrast[[idx]], 0)
-  expect_gte(fit$penalty[[idx]], 0)
+  idx <- which.max(fit_iterative$ave_sil_widths)
+  expect_gte(fit_iterative$contrast[[idx]], 0)
+  expect_gte(fit_iterative$penalty[[idx]], 0)
+  idx <- which.max(fit_var_proj$ave_sil_widths)
+  expect_gte(fit_var_proj$contrast[[idx]], 0)
[…]
