### 2020-02-02 b0005e0
Adding tests for scPCA.R
