+++ R/scPCA.R
@@ -173,6 +173,7 @@ scPCA <- function(target, background, center = TRUE, scale = FALSE,
       center = center,
       scale = scale,
       n_eigen = n_eigen,
+      alg = alg,
       contrasts = contrasts,
       penalties = penalties,
       clust_method = clust_method,
@@ -213,6 +214,7 @@ scPCA <- function(target, background, center = TRUE, scale = FALSE,
       center = center,
       scale = scale,
       n_eigen = n_eigen,
+      alg = alg,
       contrasts = cv_opt_params$contrast,
       penalties = cv_opt_params$penalty,
       clust_method = clust_method,
