### 2020-02-02 f17a717
Modified documents in spcaWrapper to avoid namescpace issues
