+++ NAMESPACE
@@ -11,7 +11,6 @@ importFrom(cluster,pam)
 importFrom(cluster,silhouette)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,full_join)
-importFrom(elasticnet,spca)
 importFrom(kernlab,as.kernelMatrix)
 importFrom(kernlab,specc)
 importFrom(matrixStats,colSds)
@@ -21,8 +20,6 @@ importFrom(origami,make_folds)
 importFrom(origami,training)
 importFrom(origami,validation)
 importFrom(purrr,reduce)
-importFrom(sparsepca,rspca)
-importFrom(sparsepca,spca)
 importFrom(stats,cov)
 importFrom(stats,cutree)
 importFrom(stats,dist)
