+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: scPCA
 Title: Sparse Contrastive Principal Component Analysis
-Version: 1.1.6
+Version: 1.3.0
 Authors@R: c(
     person(given = "Philippe",
            family = "Boileau",
@@ -36,6 +36,7 @@ Imports:
     matrixStats,
     BiocParallel,
     elasticnet,
+    sparsepca,
     cluster,
     kernlab,
     origami
