### 2020-02-02 06fa579
Updated vignette with microbenchmark for SPCA algorightms
