+++ NEWS.md
@@ -1,2 +1,17 @@
 # Changes in version 0.99.0 (2019-09-13)
 + Submitted to Bioconductor
+
+# Changes in version 1.1.2 (2020-01-08)
++ Added hierarchical clustering options for clustering based cross-validation
+
+# Changes in version 1.1.5 (2020-01-18)
++ Fixed issue with matrix normalization
++ Misc. big fixes
++ Improvements to code coverage
+
+# Changes in version 1.3.0 (2020-02-02)
++ Added more SPCA algorithm options
+  - SPCA via variable projection
+  - Randomized SPCA via variable projection
++ New vignette section comparing performance of SPCA algorithms
++ Improvements to code coverage
