+++ README.Rmd
@@ -42,18 +42,19 @@ often subject to batch effects and other technical noise. In tandem with the
 development of sequencing technology (e.g., RNA-seq, scRNA-seq), many variants
 of PCA have been developed in attempts to remedy deficiencies in
 interpretability and stability that plague their classical variants. Such
-developments have included both sparse PCA (SPCA) [@zou2006sparse], which
-increases the stability and interpretability of principal component loadings in
-high dimensions, and, more recently, contrastive PCA (cPCA)
-[@abid2018exploring], which captures relevant information in the target
-(experimental) data set by eliminating technical noise through comparison to a
-so-called background data set. While SPCA and cPCA have both individually proven
-useful in resolving distinct shortcomings of PCA, neither is capable of
-simultaneously tackling the issues of interpretability, stability and relevance
-simultaneously. The `scPCA` package implements _sparse constrastive PCA_, to
-accomplish both of these tasks in the context of high-dimensional biological
-data. In addition to implementing this newly developed technique, the `scPCA`
-package also implements cPCA and generalizations thereof.
+developments have included both sparse PCA (SPCA)
+[@zou2006sparse; @erichson2018sparse], which increases the stability and
+interpretability of principal component loadings in high dimensions, and, more
+recently, contrastive PCA (cPCA) [@abid2018exploring], which captures relevant
+information in the target (experimental) data set by eliminating technical noise
+through comparison to a so-called background data set. While SPCA and cPCA have
+both individually proven useful in resolving distinct shortcomings of PCA,
+neither is capable of simultaneously tackling the issues of interpretability,
+stability and relevance simultaneously. The `scPCA` package implements
+_sparse constrastive PCA_, to accomplish both of these tasks in the context of
+high-dimensional biological data. In addition to implementing this newly
+developed technique, the `scPCA` package also implements cPCA and
+generalizations thereof.
 
 ---
 
