+++ DESCRIPTION
@@ -50,7 +50,8 @@ Suggests:
     ggplot2,
     ggpubr,
     splatter,
-    SingleCellExperiment
+    SingleCellExperiment,
+    microbenchmark
 License: MIT + file LICENSE
 URL: https://github.com/PhilBoileau/scPCA
 BugReports: https://github.com/PhilBoileau/scPCA/issues
