### 2020-02-03 6808ac3
Revert "Fixing version numbering on devel"
This reverts commit 62ebdf8338d23f9e98193a45298bded82e4e19b2. error in version number.
