### 2020-02-03 a14bbf9
Revert "Fixing version number for bioconductor"
This reverts commit d35f7472e4562bed4c4f8fbfd83ba30d00765092. Error in versioniong.
