### 2020-02-03 8c9c649
Adding magick dependency to travi-ci configuration
