### 2020-02-03 81691ea
adding support for ImageMagick
