+++ .travis.yml
@@ -22,6 +22,24 @@ r:
 before_install:
   - Rscript -e 'BiocManager::install(version = "devel", ask = FALSE)'
 
+addons:
+   apt:
+     update: true
+     sources:
+       - sourceline: 'ppa:opencpu/imagemagick'
+       - sourceline: 'ppa:ubuntugis/ppa'
+     packages:
+       - libmagick++-dev
+       - librsvg2-dev
+       - libwebp-dev
+       - libpoppler-cpp-dev
+       - libtesseract-dev
+       - libleptonica-dev
+       - tesseract-ocr-eng
+       - r-cran-rgdal
+       - libfftw3-dev
+       - cargo
+
 r_packages:
   - devtools
   - covr
