### 2020-02-03 7eca54b
Decrease vignette running time
