### 2020-02-04 664e6d1
adding Matrix package to travis-ci
