### 2020-02-04 eb73316
Updating test to handle travis bug
