### 2020-02-04 #2055 @csoneson
👋 @PhilBoileau - I will handle your submission. While I look for suitable reviewers, could you add DOIs to the references that are currently missing them, and look into case protection to get the right capitalization of the references (e.g., r, bioconductor, some of the journal names)? To generate a new pdf proof, just type `@whedon generate pdf` in this issue.
