### 2020-02-04 #2055 @csoneson
👋 @LTLA, @fabian-s, @gabrielodom - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

`scPCA: A toolbox for sparse contrastive principal component analysis in R`
Software: https://github.com/PhilBoileau/scPCA
Paper: https://github.com/openjournals/joss-papers/blob/joss.02055/joss.02055/10.21105.joss.02055.pdf

The guidelines for JOSS reviews are available at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, and the review criteria are listed at https://joss.readthedocs.io/en/latest/review_criteria.html.

Thanks!
