### 2020-02-04 10ceecc
coerceMatrix steps now skip on Travis CI
