### 2020-02-04 442ce28
Updated references; included DOIs, fixed capitalization of titles
