+++ manuscript/paper.bib
@@ -1,14 +1,13 @@
-@article{lun2018singlecellexperiment,
-  title={SingleCellExperiment: S4 classes for single cell data},
-  author={Lun, Aaron and Risso, Davide and Korthauer, K},
-  journal={R package version},
-  volume={1},
-  number={0},
-  year={2018}
-}
+@manual{lun2018singlecellexperiment,
+    title = {SingleCellExperiment: S4 classes for single cell data},
+    author = {{Lun}, Aaron and {Risso}, Davide},
+    year = {2019},
+    note = {R package version 1.6.0},
+    doi = {10.18129/B9.bioc.SingleCellExperiment}
+  }
 
 @manual{R,
-  title = {R: A Language and Environment for Statistical Computing},
+  title = {R: A language and environment for statistical computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
@@ -18,7 +17,7 @@
 
 @article{Johnstone2009,
 author = {{Johnstone}, I.~M. and {Lu}, A.~Y. },
-title = {On Consistency and Sparsity for Principal Components Analysis in High Dimensions},
+title = {On consistency and sparsity for principal components analysis in high dimensions},
 journal = {Journal of the American Statistical Association},
 volume = {104},
 number = {486},
@@ -39,7 +38,7 @@ eprint = {
 
 @article{Zou2006,
 author = {{Zou}, H. and {Hastie}, T. and {Tibshirani}, R.},
-title = {Sparse Principal Component Analysis},
+title = {Sparse principal component analysis},
 journal = {Journal of Computational and Graphical Statistics},
 volume = {15},
 number = {2},
@@ -60,7 +59,6 @@ eprint = {
 @article{Abid2018,
 author = {{Abid}, A. and {Zhang}, M.~J. and {Bagaria}, V.~K. and {Zou}, J.},
 doi = {10.1038/s41467-018-04608-8},
-file = {:home/phil/.local/share/data/Mendeley Ltd./Mendeley Desktop/Downloaded/Abid et al. - 2018 - Exploring patterns enriched in a dataset with contrastive principal component analysis.pdf:pdf},
 issn = {2041-1723},
 journal = {Nature Communications},
 keywords = {Computational biology and bioinformatics,Machine learning,Statistics},
@@ -76,7 +74,7 @@ year = {2018}
 
 @article {Boileau,
 	author = {{Boileau}, P. and {Hejazi}, N.~S. and {Dudoit}, S.},
-	title = {Exploring High-Dimensional Biological Data with Sparse Contrastive Principal Component Analysis},
+	title = {Exploring high-dimensional biological data with sparse contrastive principal component analysis},
 	year = {2019},
 	doi = {10.1101/836650},
 	publisher = {Cold Spring Harbor Laboratory},
@@ -86,30 +84,33 @@ year = {2018}
 }
 
 @article{gentleman2004bioconductor,
-  title={Bioconductor: open software development for computational biology and bioinformatics},
-  author={{Gentleman}, R. and {Carey}, V.~J. and {Bates}, D.~M. and {Bolstad}, B. and {Dettling}, M. and {Dudoit}, S. and {Ellis}, B. and {Gautier}, L. and {Ge}, Y. and {Gentry}, J. and {others}},
-  journal={Genome biology},
-  volume={5},
-  number={10},
-  pages={R80},
-  year={2004},
-  publisher={BioMed Central}
+  title = {Bioconductor: open software development for computational biology and bioinformatics},
[…]
