### 2020-02-04 #2055 @PhilBoileau
Thank you, @csoneson! I have added the doi's, and fixed the issues with journal name/article title capitalization.
