### 2020-02-04 4f416ef
Fixed capitalization in bib
