+++ manuscript/paper.bib
@@ -96,7 +96,7 @@ year = {2018}
 }
 
 @book{gentleman2006bioinformatics,
-  title={Bioinformatics and computational biology solutions using R and Bioconductor},
+  title={Bioinformatics and computational biology solutions using {R} and {B}ioconductor},
   author={{Gentleman}, R. and {Carey}, V. and {Huber}, W. and {Irizarry}, R. and {Dudoit}, S.},
   year={2006},
   publisher={Springer Science \& Business Media},
@@ -104,7 +104,7 @@ year = {2018}
 }
 
 @article{huber2015orchestrating,
-  title = {Orchestrating high-throughput genomic analysis with Bioconductor},
+  title = {Orchestrating high-throughput genomic analysis with {B}ioconductor},
   author = {{Huber}, W. and {Carey}, V.~J. and {Gentleman}, R. and {Anders}, S. and {Carlson}, M. and {Carvalho}, B.~S. and {Bravo}, H.~C. and {Davis}, S. and {Gatto}, L. and {Girke}, T. and {others}},
   journal = {Nature Methods},
   volume = {12},
