### 2020-02-04 dfedfa9
skipping unit test on matrix conversion, need to figure out why failing on linux and windows
