### 2020-02-04 f18a290
let BiocManager deal with versioning
