+++ appveyor.yml
@@ -1,4 +1,15 @@
+#----------------------------------------------------------------
+# AppVeyor configuration for R packages
+#
+# REFERENCES:
+# * AppVeyor CI: https://ci.appveyor.com/
+# * r-appveyor: https://github.com/krlmlr/r-appveyor
+#
+# Validate your .appveyor.yml file at
+# https://ci.appveyor.com/tools/validate-yaml
+
 # DO NOT CHANGE the "init" and "install" sections below
+#----------------------------------------------------------------
 
 # Download script file from GitHub
 init:
@@ -7,12 +18,14 @@ init:
         Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
         Import-Module '..\appveyor-tool.ps1'
 install:
-  ps: Bootstrap
+  ps: |
+    Bootstrap
+    InstallRtools
 
+# Adapt as necessary starting from here
 cache:
   - C:\RLibrary -> appveyor.yml
 
-# Adapt as necessary starting from here
 branches:
   only:
     - master
@@ -25,18 +38,20 @@ environment:
     PKGTYPE: both
 
   matrix:
-    - R_VERSION: release
+    - R_VERSION: bioc-release
+      R_ARCH: x64
+
+    - R_VERSION: bioc-devel
       R_ARCH: x64
 
 build_script:
   - echo Current directory=%CD%
   - travis-tool.sh install_r base64
+  - travis-tool.sh install_r BiocManager testthat knitr
   - travis-tool.sh install_github r-lib/remotes r-lib/covr
-  - travis-tool.sh install_r BiocManager testthat knitr elasticnet sparsepca
-  - travis-tool.sh install_bioc BiocStyle BiocCheck BiocGenerics
-  - travis-tool.sh install_bioc S4Vectors splatter SingleCellExperiment
-  - travis-tool.sh install_bioc_deps
-  - travis-tool.sh install_deps
+  - travis-tool.sh install_bioc BiocStyle BiocCheck BiocGenerics S4Vectors
+  - travis-tool.sh install_r matrixStats elasticnet sparsepca
+  - travis-tool.sh install_bioc splatter SingleCellExperiment
 
 test_script:
   - travis-tool.sh run_tests
