+++ .travis.yml
@@ -20,7 +20,7 @@ r:
   - bioc-devel
 
 before_install:
-  - Rscript -e 'BiocManager::install(version = "devel", ask = FALSE)'
+  - Rscript -e 'BiocManager::install(ask = FALSE)'
 
 addons:
    apt:
@@ -48,6 +48,7 @@ r_packages:
   - sparsepca
   - magick
   - Matrix
+  - matrixStats
 
 r_github_packages:
   - r-lib/sessioninfo
