### 2020-02-04 543a603
switch appveyor to devel; version bump
