+++ appveyor.yml
@@ -20,7 +20,6 @@ init:
 install:
   ps: |
     Bootstrap
-    InstallRtools
 
 # Adapt as necessary starting from here
 cache:
