### 2020-02-07 #2079 @whedon
[REVIEW]: scPCA: A toolbox for sparse contrastive principal component analysis in R
**Submitting author:** @PhilBoileau (<a href="http://orcid.org/0000-0002-4850-2507">Philippe Boileau</a>)
**Repository:** <a href="https://github.com/PhilBoileau/scPCA" target ="_blank">https://github.com/PhilBoileau/scPCA</a>
**Version:** v1.1.10
