### 2020-02-07 #2079 @PhilBoileau
@csoneson - Thank you for organizing the review process! I reviewed the checklist, the `scPCA` package satisfies all the requirements.

@fabian-s, @LTLA - Thank you for volunteering to review our package. We're looking forward to receiving your feedback.
