### 2020-02-10 PhilBoileau/scPCA#36 @fabian-s
JOSS: "Do the authors describe how this software compares to other commonly-used packages?"
I don't think you do - maybe add a remark that commonly used non-linear dimension reduction methods for single cell experiments like UMAP,  t-SNE, Isomap etc don't provide *interpretable* results?
