### 2020-02-10 #2079 @fabian-s
This seems fine -- one minor nitpick at PhilBoileau/scPCA/issues/36 regarding a potentially missing comparison to existing alternative approaches, but that is debatable...
