### 2020-02-11 0f343db
mention non-linear methods in scPCA manuscript
