+++ manuscript/paper.bib
@@ -1,10 +1,10 @@
 @manual{lun2018singlecellexperiment,
-    title = {SingleCellExperiment: S4 classes for single cell data},
-    author = {{Lun}, Aaron and {Risso}, Davide},
-    year = {2019},
-    note = {R package version 1.6.0},
-    doi = {10.18129/B9.bioc.SingleCellExperiment}
-  }
+  title = {SingleCellExperiment: S4 classes for single cell data},
+  author = {{Lun}, Aaron and {Risso}, Davide},
+  year = {2019},
+  note = {R package version 1.6.0},
+  doi = {10.18129/B9.bioc.SingleCellExperiment}
+}
 
 @manual{R,
   title = {R: A language and environment for statistical computing},
@@ -16,60 +16,60 @@
 }
 
 @article{Johnstone2009,
-author = {{Johnstone}, I.~M. and {Lu}, A.~Y. },
-title = {On consistency and sparsity for principal components analysis in high dimensions},
-journal = {Journal of the American Statistical Association},
-volume = {104},
-number = {486},
-pages = {682-693},
-year  = {2009},
-publisher = {Taylor & Francis},
-doi = {10.1198/jasa.2009.0121},
-    note ={PMID: 20617121},
+  author = {{Johnstone}, I.~M. and {Lu}, A.~Y. },
+  title = {On consistency and sparsity for principal components analysis in high dimensions},
+  journal = {Journal of the American Statistical Association},
+  volume = {104},
+  number = {486},
+  pages = {682-693},
+  year  = {2009},
+  publisher = {Taylor & Francis},
+  doi = {10.1198/jasa.2009.0121},
+      note ={PMID: 20617121},
 
-URL = { 
-        https://doi.org/10.1198/jasa.2009.0121
-    
-},
-eprint = { 
-        https://doi.org/10.1198/jasa.2009.0121   
-}
+  URL = { 
+          https://doi.org/10.1198/jasa.2009.0121
+      
+  },
+  eprint = { 
+          https://doi.org/10.1198/jasa.2009.0121   
+  }
 }
 
 @article{Zou2006,
-author = {{Zou}, H. and {Hastie}, T. and {Tibshirani}, R.},
-title = {Sparse principal component analysis},
-journal = {Journal of Computational and Graphical Statistics},
-volume = {15},
-number = {2},
-pages = {265-286},
-year  = {2006},
-publisher = {Taylor & Francis},
-doi = {10.1198/106186006X113430},
+  author = {{Zou}, H. and {Hastie}, T. and {Tibshirani}, R.},
+  title = {Sparse principal component analysis},
+  journal = {Journal of Computational and Graphical Statistics},
+  volume = {15},
+  number = {2},
+  pages = {265-286},
+  year  = {2006},
+  publisher = {Taylor & Francis},
+  doi = {10.1198/106186006X113430},
 
-URL = {
-        https://doi.org/10.1198/106186006X113430
+  URL = {
+          https://doi.org/10.1198/106186006X113430
 
-},
-eprint = {
-        https://doi.org/10.1198/106186006X113430
-}
+  },
+  eprint = {
+          https://doi.org/10.1198/106186006X113430
+  }
 }
 
 @article{Abid2018,
-author = {{Abid}, A. and {Zhang}, M.~J. and {Bagaria}, V.~K. and {Zou}, J.},
-doi = {10.1038/s41467-018-04608-8},
-issn = {2041-1723},
-journal = {Nature Communications},
-keywords = {Computational biology and bioinformatics,Machine learning,Statistics},
-month = {dec},
-number = {1},
-pages = {2134},
-publisher = {Nature Publishing Group},
[…]
