### 2020-02-11 40633f7
fixed issue with zinbwave author list
