+++ manuscript/paper.bib
@@ -124,7 +124,8 @@
   publisher = {Nature Publishing Group},
   title = {{A general and flexible method for signal extraction from single-cell {RNA}-seq data}},
   volume = {9},
-  year = {2018}
+  year = {2018},
+  author = {{Risso}, D. and {Perraudeau}, F. and {Dudoit}, S. and {Vert}, J.-P.}
 }
 
 @misc{lel2018umap,
