### 2020-02-11 #2079 @PhilBoileau
Thanks, @fabian-s. I've added a line that briefly details the results of our comparisons to existing methods.
