### 2020-02-11 863f261
edited comparison to other methods
