+++ manuscript/paper.bib
@@ -115,19 +115,6 @@
   doi = {10.1038/nmeth.3252}
 }
 
-@article{risso2018zinb,
-  doi = {10.1038/s41467-017-02554-5},
-  issn = {20411723},
-  journal = {Nature Communications},
-  month = {dec},
-  number = {1},
-  publisher = {Nature Publishing Group},
-  title = {{A general and flexible method for signal extraction from single-cell {RNA}-seq data}},
-  volume = {9},
-  year = {2018},
-  author = {{Risso}, D. and {Perraudeau}, F. and {Dudoit}, S. and {Vert}, J.-P.}
-}
-
 @misc{lel2018umap,
   title={UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction},
   author={{McInnes}, L. and {Healy}, J. and {Melville}, J.s},
@@ -135,20 +122,6 @@
   eprint={1802.03426},
   archivePrefix={arXiv},
   primaryClass={stat.ML}
-}
-
-@article{Wang2017,
-  author = {{Wang}, B. and {Zhu}, J. and {Pierson}, E. and {Ramazzotti}, D. and {Batzoglou}, S.},
-  doi = {10.1038/nMeth.4207},
-  issn = {15487105},
-  journal = {Nature Methods},
-  month = {mar},
-  number = {4},
-  pages = {414--416},
-  publisher = {Nature Publishing Group},
-  title = {{Visualization and analysis of single-cell {RNA}-seq data by kernel-based similarity learning}},
-  volume = {14},
-  year = {2017}
 }
 
 @article{vanDerMaaten2008,
