### 2020-02-11 #2079 @PhilBoileau
I made slight edits to the newly added sentence, should read a little better now.
