### 2020-02-14 8d996ee
adding JOSS badge to README
