### 2020-02-22 PhilBoileau/scPCA#38 @LTLA
JOSS: cut down on the package load chat
Package load gives me:

```
> library(scPCA)
scPCA v1.1.9: Sparse Contrastive Principal Component Analysis
```

I'm sure that anyone loading **scPCA** explicitly knows what they're in for.

See also comments at https://stat.ethz.ch/pipermail/bioc-devel/2019-April/014967.html.
