### 2020-02-22 PhilBoileau/scPCA#39 @LTLA
JOSS: hidden package imports make it difficult to copy-paste vignette code
The `spca` function is the worst offender, with no indication of what package it came from.
