### 2020-02-22 PhilBoileau/scPCA#40 @LTLA
JOSS: natively handling alternative matrix representations
It looks like `scPCA::coerceMatrix` collapses sparse matrices to an ordinary matrix. This isn't healthy for memory efficiency or speed when zeroes are unnecessarily stored and processed.

I'm not going to make a big deal about this because it seems that your upstream packages coerce to dense matrices anyway, so it doesn't really make much difference whether you do the right thing for these alternative matrices. However, the proposed use case for scRNA-seq data will not be a strong one if people have to desparsify matrices. A small dataset of 50K cells, 2K HVGs requires almost 1 GB of memory in theory, and practical usage will be several-fold higher due to all the copies.

You may consider suggesting to the maintainers of your upstream packages to make adjustments to accept sparse matrices (or, in fact, matrices of any representation). e.g., erichson/rSVD#4.

While I'm on this point, your coercion doesn't catch other matrix classes, e.g., `HDF5Matrix` and `DelayedMatrix` objects will just slip through and presumably cause havoc in later parts of the code.
