### 2020-02-22 PhilBoileau/scPCA#41 @LTLA
JOSS: practical generation of the background in scRNA-seq studies
The vignette has a section describing how **scPCA** might be used for scRNA-seq data analysis. This is all well and good but it splits cells into target and background based on biological groupings that will generally not be _a priori_ available to analysts.

It would be worth mentioning how you expect users to do this split on their data. Do an initial analysis to identify putative cell types and then use those as the groupings? (Seems sort of circular to me.) Or perhaps rely on some pre-defined experimental design with cell controls?
