### 2020-02-22 PhilBoileau/scPCA#42 @LTLA
JOSS: small scRNA-seq test dataset takes... quite a while.
```r
library(scRNAseq)
sce <- LunSpikeInData()
sce2 <- LunSpikeInData("tropho")

# Skipping all the QC for simplicity.
library(scater)
sce <- logNormCounts(sce)
sce2 <- logNormCounts(sce2)

library(scran)
dec <- modelGeneVar(sce)
hvgs <- getTopHVGs(dec, n=1000)
hvgs <- intersect(hvgs, rownames(sce2))

library(scPCA)
stuff <- t(logcounts(sce)[hvgs,])
stuff2 <- t(logcounts(sce2)[hvgs,])
out <- scPCA(stuff, stuff2, alg="rand_var_proj", n_centers=5)
```

It's been running for some time (20, 30 minutes?), and there's only a total of 384 cells here.

<summary>Session information</summary>

```
R Under development (unstable) (2020-02-05 r77773)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 18.04.4 LTS

Matrix products: default
BLAS:   /home/luna/Software/R/trunk/lib/libRblas.so
LAPACK: /home/luna/Software/R/trunk/lib/libRlapack.so

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C

attached base packages:
[1] parallel  stats4    stats     graphics  grDevices utils     datasets
[8] methods   base

other attached packages:
 [1] scPCA_1.1.9                 scran_1.15.21
 [3] scater_1.15.26              ggplot2_3.2.1
 [5] ensembldb_2.11.3            AnnotationFilter_1.11.0
 [7] GenomicFeatures_1.39.5      AnnotationDbi_1.49.1
 [9] scRNAseq_2.1.6              SingleCellExperiment_1.9.2
[11] SummarizedExperiment_1.17.2 DelayedArray_0.13.4
[13] BiocParallel_1.21.2         matrixStats_0.55.0
[15] Biobase_2.47.2              GenomicRanges_1.39.2
[17] GenomeInfoDb_1.23.13        IRanges_2.21.3
[19] S4Vectors_0.25.12           BiocGenerics_0.33.0

  [1] ggbeeswarm_0.6.0              colorspace_1.4-1
  [3] XVector_0.27.0                BiocNeighbors_1.5.1
  [5] sparsepca_0.1.2               listenv_0.8.0
  [7] bit64_0.9-7                   interactiveDisplayBase_1.25.0
  [9] codetools_0.2-16              Rsamtools_2.3.4
 [11] cluster_2.1.0                 kernlab_0.9-29
 [13] dbplyr_1.4.2                  shiny_1.4.0
 [15] BiocManager_1.30.10           compiler_4.0.0
 [17] httr_1.4.1                    dqrng_0.2.1
 [19] assertthat_0.2.1              Matrix_1.2-18
 [21] fastmap_1.0.1                 lazyeval_0.2.2
 [23] limma_3.43.4                  later_1.0.0
 [25] BiocSingular_1.3.2            htmltools_0.4.0
 [27] prettyunits_1.1.1             tools_4.0.0
 [29] rsvd_1.0.3                    igraph_1.2.4.2
 [31] gtable_0.3.0                  glue_1.3.1
 [33] GenomeInfoDbData_1.2.2        dplyr_0.8.4
 [35] rappdirs_0.3.1                Rcpp_1.0.3
 [37] vctrs_0.2.3                   Biostrings_2.55.4
 [39] ExperimentHub_1.13.5          rtracklayer_1.47.0
 [41] DelayedMatrixStats_1.9.0      gbRd_0.4-11
 [43] origami_1.0.3                 stringr_1.4.0
 [45] globals_0.12.5                mime_0.9
 [47] lifecycle_0.1.0               irlba_2.3.3
 [49] statmod_1.4.34                XML_3.99-0.3
 [51] future_1.16.0                 AnnotationHub_2.19.7
 [53] edgeR_3.29.0                  zlibbioc_1.33.1
 [55] scales_1.1.0                  hms_0.5.3
 [57] promises_1.1.0                ProtGenerics_1.19.3
 [59] yaml_2.2.1                    curl_4.3
 [61] memoise_1.1.0                 gridExtra_2.3
 [63] biomaRt_2.43.3                stringi_1.4.6
 [65] RSQLite_2.2.0                 BiocVersion_3.11.1
 [67] bibtex_0.4.2.2                Rdpack_0.11-1
 [69] rlang_0.4.4                   pkgconfig_2.0.3
 [71] bitops_1.0-6                  lattice_0.20-40
 [73] purrr_0.3.3                   GenomicAlignments_1.23.1
 [75] bit_1.1-15.2                  tidyselect_1.0.0
 [77] magrittr_1.5                  R6_2.4.1
 [79] DBI_1.1.0                     pillar_1.4.3
 [81] withr_2.1.2                   abind_1.4-5
 [83] RCurl_1.98-1.1                future.apply_1.4.0
 [85] tibble_2.1.3                  crayon_1.3.4
 [87] BiocFileCache_1.11.4          viridis_0.5.1
 [89] progress_1.2.2                locfit_1.5-9.1
 [91] grid_4.0.0                    data.table_1.12.8
 [93] blob_1.2.1                    digest_0.6.24
 [95] xtable_1.8-4                  httpuv_1.5.2
 [97] openssl_1.4.1                 munsell_0.5.0
 [99] beeswarm_0.2.3                viridisLite_0.3.0
[101] vipor_0.4.5                   askpass_1.1
```
