### 2020-02-22 #2079 @LTLA
Looks good in general. Various comments:

- PhilBoileau/scPCA#38
- PhilBoileau/scPCA#39
- PhilBoileau/scPCA#40
- PhilBoileau/scPCA#41
- PhilBoileau/scPCA#42
