### 2020-02-22 b54d38a
explicitely show vignette's package imports in section 3
