+++ vignettes/scpca_intro.Rmd
@@ -18,12 +18,12 @@ knitr::opts_chunk$set(echo = TRUE)
 ```
 
 Data pre-processing and exploratory data analysis and are two important steps in
-the data science life-cycle. As data sets become larger and the signal weaker,
+the data science life-cycle. As datasets become larger and the signal weaker,
 their importance increases. Thus, methods that are capable of extracting the
-signal from such data sets is badly needed. Often, these steps rely on
+signal from such datasets is badly needed. Often, these steps rely on
 dimensionality reduction techniques to isolate pertinent information in data.
 However, many of the most commonly-used methods fail to reduce the dimensions of
-these large and noisy data sets successfully.
+these large and noisy datasets successfully.
 
 Principal component analysis (PCA) is one such method. Although popular for its
 interpretable results and ease of implementation, PCA's performance on
@@ -35,7 +35,8 @@ Fortunately, modifications of PCA have been developed to remedy these issues.
 Namely, sparse PCA (sPCA) was created to increase the stability of the principal
 component loadings and variable scores in high dimensions, and constrastive PCA
 (cPCA) was proposed as a method for capturing relevant information in the
-high-dimensional data [@abid2018exploring].
+high-dimensional data by harnessing variation in control data
+[@abid2018exploring].
 
 Although sPCA and cPCA have proven useful in resolving individual shortcomings
 of PCA, neither is capable of tackling the issues of stability and relevance
@@ -83,8 +84,8 @@ A brief comparison of PCA, SPCA, cPCA and scPCA is provided below. All four
 methods are applied to a simulated target dataset consisting of `r nrow(toy_df)`
 observations and `r ncol(toy_df) - 1` continuous variables. Additionally, each
 observation is classified as belonging to one of four classes. This label is
-known a priori. A background data set is comprised of the same number of
-variables as the target data set.
+known a priori. A background dataset is comprised of the same number of
+variables as the target dataset, representing control data.
 
 The target data was simulated as follows:
 
@@ -164,7 +165,7 @@ p_spca
 ## Contrastive PCA (cPCA)
 
 The first two contrastive principal components of cPCA successfully captured
-the variation of interest in the data with the help of the background data set.
+the variation of interest in the data with the help of the background dataset.
 To fit contrastive PCA with the `scPCA` function of this package, simply select
 no penalization (by setting argument `penalties = 0`), and specify the expected
 number of clusters in the data. Here, we set the number of clusters to 4
@@ -339,7 +340,7 @@ The `scPCA` package provides three options with which to sparsify the loadings
 produced by cPCA:
 1. The traditional iterative SPCA algorithm by @zou2006sparse, implemented in the `elasticnet` package.
 2. The SPCA algorithm relying on variable projection by @erichson2018sparse, implemented in the `sparsepca` package.
-3. The randommized SPCA algorithm, which uses variable projection and random numerical linear algebra methods, by @erichson2018sparse, implemented in the `sparsepca` package.
+3. The randomized SPCA algorithm, which uses variable projection and random numerical linear algebra methods, by @erichson2018sparse, implemented in the `sparsepca` package.
 
 For historical reasons, the default SPCA algorithm used is that of
 @zou2006sparse. However, @erichson2018sparse's methods are noticeably faster.
@@ -378,7 +379,7 @@ non-randomized counterpart [@erichson2018sparse].
 We now turn to discussing how the tools in the `scPCA` package can be used more
 readily with data structures common in computational biology by examining their
 integration with the `SingleCellExperiment` container class. For our example, we
-will use `splatter` to simulate a scRNA-seq data set using the Splatter
+will use `splatter` to simulate a scRNA-seq dataset using the Splatter
 framework [@zappia2017splatter]. This method simulates a scRNA-seq count matrix
[…]
