### 2020-02-22 48a305d
touching up JOSS paper
