### 2020-02-22 cb1879c
Added blurb on efficiency of scPCA on large datasets
