+++ inst/REFERENCES.bib
@@ -51,3 +51,17 @@
 	eprint = {https://www.biorxiv.org/content/early/2019/11/09/836650.full.pdf},
 	journal = {bioRxiv}
 }
+
+@article{vanDerMaaten2008,
+author = {van der Maaten, Laurens and Hinton, Geoffrey},
+biburl = {https://www.bibsonomy.org/bibtex/28b9aebb404ad4a4c6a436ea413550b30/lopusz_kdd},
+interhash = {370ba8b9e1909b61880a6f47c93bcd49},
+intrahash = {8b9aebb404ad4a4c6a436ea413550b30},
+journal = {Journal of Machine Learning Research},
+keywords = {dimensionality_reduction tSNE visualization},
+pages = {2579--2605},
+title = {Visualizing Data using {t-SNE} },
+url = {http://www.jmlr.org/papers/v9/vandermaaten08a.html},
+volume = 9,
+year = 2008
+}
