### 2020-02-22 PhilBoileau/scPCA#40 @PhilBoileau
Thank you for pointing this out. That is definitely something that I will get on top of in the near future!

With respect to the `coerceMatrix` function not catching other matrix classes, there is a separate function, `checkArgs`, that is called at the beginning of the `scPCA ` function which performs checks on most of `scPCA`'s inputs. In particular, the target and background datasets' classes are inspected, and an error is thrown if something is off.
