+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: scPCA
 Title: Sparse Contrastive Principal Component Analysis
-Version: 1.1.9
+Version: 1.1.10
 Authors@R: c(
     person(given = "Philippe",
            family = "Boileau",
