### 2020-02-22 #2079 @PhilBoileau
Thank you for your thorough review, @LTLA! I've addressed most of your comments in a recent PR: [PhilBoileau/scPCA/pull/43](https://github.com/PhilBoileau/scPCA/pull/43). I've also responded to [PhilBoileau/scPCA#40](https://github.com/PhilBoileau/scPCA/issues/40). I hope that you find the changes satisfactory!
