### 2020-02-24 #2079 @csoneson
@PhilBoileau - this looks all good! Next steps:

- [ ] Make a tagged release of the package, and post the version tag of the release here
- [ ] Create a Zenodo archive. Make sure that the Zenodo record has the same title and author list as the paper (you can also add ORCiDs)
- [ ] Post the doi of the Zenodo record here

I would also recommend clarifying to users (e.g. in your README) the relationship between this paper and the bioRxiv preprint, e.g. to make it easier to know when to cite which paper.
