### 2020-02-24 f6554f9
Updated readme with citation section
