### 2020-02-24 #2079 @PhilBoileau
Thanks, @csoneson! Here you go:

- The release version tag is [v1.1.10](https://github.com/PhilBoileau/scPCA/releases/tag/v1.1.10)
- The Zenodo doi is [10.5281/zenodo.3686201](https://doi.org/10.5281/zenodo.3686201))

I've also added a *Citation* section in the README indicating how to cite the package. I'll update it to include the JOSS reference once published.
