+++ README.md
@@ -17,6 +17,7 @@ status](http://www.bioconductor.org/shields/build/release/bioc/scPCA.svg)](https
 [![Bioc
 Time](http://bioconductor.org/shields/years-in-bioc/scPCA.svg)](https://bioconductor.org/packages/release/bioc/html/scPCA.html)
 [![status](https://joss.theoj.org/papers/7f0f1271ede7aba120d71c9b5a14c865/status.svg)](https://joss.theoj.org/papers/7f0f1271ede7aba120d71c9b5a14c865)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3686201.svg)](https://doi.org/10.5281/zenodo.3686201)
 [![MIT
 license](http://img.shields.io/badge/license-MIT-brightgreen.svg)](http://opensource.org/licenses/MIT)
 
