### 2020-02-25 #2079 @labarba
Big thanks to our editor: @csoneson, and the reviewers: @fabian-s, @LTLA — your contribution to JOSS is appreciated 🙏
